// src/models/account-stat.model.ts
export interface AccountStatModel {
  id: string;
  bossKilled: number;
  heroDied: number;
  dungeonCompleted: number;
  chickedUsed: number;
  itemsAugment: number;
  successfulAugment: number;
  maxAugmentLevel: number;
  itemsRepaired: number;
  itemsUsed: number;
  pvpFights: number;
  pvpWins: number;
}